+++ Pandora/Paradigm/Primary/Transformer/Backwards.hs
@@ -4,7 +4,7 @@ import Pandora.Pattern.Semigroupoid ((.))
 import Pandora.Pattern.Category (($), (#))
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)))
 import Pandora.Pattern.Functor.Contravariant (Contravariant ((>$<)))
-import Pandora.Pattern.Functor.Semimonoidal (Semimonoidal (multiply))
+import Pandora.Pattern.Functor.Semimonoidal (Semimonoidal (mult))
 import Pandora.Pattern.Functor.Monoidal (Monoidal (unit))
 import Pandora.Pattern.Functor.Traversable (Traversable ((<<-)))
 import Pandora.Pattern.Functor.Distributive (Distributive ((-<<)))
@@ -27,14 +27,14 @@ instance Covariant (->) (->) t => Covariant (->) (->) (Backwards t) where
 
 -- TODO: check that effects evaluation goes in opposite order
 instance (Semimonoidal (->) (:*:) (:*:) t, Covariant (->) (->) t) => Semimonoidal (->) (:*:) (:*:) (Backwards t) where
-	multiply (Backwards x :*: Backwards y) = Backwards #
+	mult (Backwards x :*: Backwards y) = Backwards #
 		((:*:) %) <$> y -<*>- x
 
 instance (Covariant (->) (->) t, Monoidal (->) (->) (:*:) (:*:) t) => Monoidal (->) (->) (:*:) (:*:) (Backwards t) where
 	unit _ f = Backwards . point $ f One
 
 instance (Semimonoidal (<--) (:*:) (:*:) t, Covariant (->) (->) t) => Semimonoidal (<--) (:*:) (:*:) (Backwards t) where
-	multiply = Flip $ (Backwards <-> Backwards) . run (multiply @(<--) @(:*:) @(:*:)) . run
+	mult = Flip $ (Backwards <-> Backwards) . run (mult @(<--) @(:*:) @(:*:)) . run
 
 instance (Covariant (->) (->) t, Monoidal (<--) (->) (:*:) (:*:) t) => Monoidal (<--) (->) (:*:) (:*:) (Backwards t) where
 	unit _ = Flip $ \(Backwards x) -> (\_ -> extract x)