+++ Pandora/Paradigm/Schemes/UT.hs
@@ -5,7 +5,7 @@ import Pandora.Pattern.Semigroupoid ((.))
 import Pandora.Pattern.Category (($), identity)
 import Pandora.Pattern.Functor.Covariant (Covariant, Covariant ((<$>)), (-<$$>-))
 import Pandora.Pattern.Functor.Contravariant (Contravariant)
-import Pandora.Pattern.Functor.Semimonoidal (Semimonoidal (multiply))
+import Pandora.Pattern.Functor.Semimonoidal (Semimonoidal (mult))
 import Pandora.Pattern.Functor.Monoidal (Monoidal (unit))
 import Pandora.Pattern.Functor.Bindable (Bindable ((=<<)))
 import Pandora.Pattern.Functor.Traversable (Traversable ((<<-)))
@@ -36,7 +36,7 @@ instance (Covariant (->) (->) t, Covariant (->) (->) u) => Covariant (->) (->) (
 	f <$> x = UT $ f -<$$>- run x
 
 instance (Covariant (->) (->) u, Semimonoidal (->) (:*:) (:*:) t, Semimonoidal (->) (:*:) (:*:) u) => Semimonoidal (->) (:*:) (:*:) (t <.:> u) where
-	multiply (UT x :*: UT y) = UT $ multiply @(->) @(:*:) <$> multiply (x :*: y)
+	mult (UT x :*: UT y) = UT $ mult @(->) @(:*:) <$> mult (x :*: y)
 
 instance (Covariant (->) (->) t, Covariant (->) (->) u, Semimonoidal (->) (:*:) (:*:) u, Monoidal (->) (->) (:*:) (:*:) t, Monoidal (->) (->) (:*:) (:*:) u) => Monoidal (->) (->) (:*:) (:*:) (t <.:> u) where
 	unit _ f = UT . point . point $ f One
@@ -45,9 +45,9 @@ instance (Traversable (->) (->) t, Bindable (->) t, Semimonoidal (->) (:*:) (:*:
 	f =<< UT x = UT $ ((identity =<<) <$>) . (run . f <<-) =<< x
 
 instance (Covariant (->) (->) u, Semimonoidal (<--) (:*:) (:*:) t, Semimonoidal (<--) (:*:) (:*:) u) => Semimonoidal (<--) (:*:) (:*:) (t <.:> u) where
-	multiply = Flip $ \(UT xys) -> (UT <-> UT)
-		. run (multiply @(<--) @(:*:) @(:*:))
-		$ run (multiply @(<--) @(:*:) @(:*:)) <$> xys
+	mult = Flip $ \(UT xys) -> (UT <-> UT)
+		. run (mult @(<--) @(:*:) @(:*:))
+		$ run (mult @(<--) @(:*:) @(:*:)) <$> xys
 
 instance (Covariant (->) (->) u, Monoidal (<--) (->) (:*:) (:*:) t, Monoidal (<--) (->) (:*:) (:*:) u) => Monoidal (<--) (->) (:*:) (:*:) (t <.:> u) where
 	unit _ = Flip $ \(UT x) -> (\_ -> extract $ extract x)