+++ Pandora/Pattern/Functor/Semimonoidal.hs
@@ -3,4 +3,4 @@ module Pandora.Pattern.Functor.Semimonoidal where
 import Pandora.Pattern.Semigroupoid (Semigroupoid)
 
 class Semigroupoid p => Semimonoidal p source target t where
-	multiply :: p (source (t a) (t b)) (t (target a b))
+	mult :: p (source (t a) (t b)) (t (target a b))