+++ CHANGELOG.md
@@ -592,3 +592,4 @@
 * Define `-->` type synonymous
 * Define `Opposite` type family
 * Define experimental `Appliable` typeclass
+* Rename `multiply` method to `mult` in `Semimonoidal` typeclass